renamed comm.sleekxmpp back to comm.xmpp
