+++ pyobs/comm/slixmpp/xmppcomm.py
@@ -42,7 +42,7 @@ class XmppComm(Comm):
     like this::
 
         comm:
-            class: pyobs.sleekxmpp.XmppComm
+            class: pyobs.xmpp.XmppComm
             jid:  someuser@example.com/pyobs
 
     Using this, *pyobs* tries to connect to example.com as user ``someuser`` with resource ``pyobs``. Since ``pyobs``
@@ -59,7 +59,7 @@ class XmppComm(Comm):
     in the same directory as the module config::
 
         comm_cfg: &comm
-            class: pyobs.comm.sleekxmpp.XmppComm
+            class: pyobs.comm.xmpp.XmppComm
             domain: example.com
 
     Now in the module configuration, one can simply do this::
@@ -84,7 +84,7 @@ class XmppComm(Comm):
     parameter must be True, and False otherwise. Cryptic error messages will follow, if one does not set this properly.
 
     """
-    __module__ = 'pyobs.comm.sleekxmpp'
+    __module__ = 'pyobs.comm.xmpp'
 
     def __init__(self, jid: Optional[str] = None, user: Optional[str] = None, domain: Optional[str] = None,
                  resource: str = 'pyobs', password: str = '', server: Optional[str] = None,
@@ -190,7 +190,7 @@ class XmppComm(Comm):
         # close parent class
         Comm.close(self)
 
-        # disconnect from sleekxmpp server
+        # disconnect from xmpp server
         self._xmpp.disconnect()
 
     def _xmpp_thread(self):
@@ -445,7 +445,7 @@ class XmppComm(Comm):
         body = json.loads(xml.sax.saxutils.unescape(msg['pubsub_event']['items']['item']['payload'].text))
 
         # do we have a <delay> element?
-        delay = msg.findall('{urn:sleekxmpp:delay}delay')
+        delay = msg.findall('{urn:xmpp:delay}delay')
         if len(delay) > 0:
             # ignore this message
             return
