+++ pyobs/comm/xmpp/__init__.py
