+++ pyobs/comm/xmpp/rpc.py
