+++ pyobs/comm/xmpp/xmppclient.py
@@ -4,8 +4,8 @@ import sleekxmpp
 import sleekxmpp.xmlstream
 from typing import List, Any
 
-from pyobs.comm.sleekxmpp.xep_0009.rpc import XEP_0009
-from pyobs.comm.sleekxmpp.xep_0009_timeout import XEP_0009_timeout
+from pyobs.comm.xmpp.xep_0009.rpc import XEP_0009
+from pyobs.comm.xmpp.xep_0009_timeout import XEP_0009_timeout
 
 
 log = logging.getLogger(__name__)
